Requirement: #4133 - Rework Duplicate Sop Policies
- Fixed connection leak in ServerCommandProcessor.

- Fixed errors that occur when the existing images no longer exist in the study folder by the time when the duplicates are processed. This  happens if the existing SOP instances contain different demographics from the rest of the study and aren't yet processed when the duplicates are received. When the existing SOP instances are processed, they are moved to the reconcile queue. Received duplicate instances are no longer duplicate when they are processed.

- Fixed errors that occurs when images are reconciled but another copy of the images is present in the study folder. This happens if two copies of the images with different demographics from the study are sent to the server at different times. Both end up in the SIQ as non-duplicate entries. After the first set is reconciled, 2nd set will become "duplicate" of the first set.

The ReconcileStudy.SaveFileCommand now throws InstanceAlreadyExistsException when this scenario happens. The file will be handled by DuplicateSopProcessor which will move the files to another folder and inserts StudyProcess workqueue entry to reprocess them. A new duplicate SIQ entry may be created afterwards.

+++ ImageServer/Common/CommandProcessor/RenameFileCommand.cs
@@ -47,9 +47,11 @@ namespace ClearCanvas.ImageServer.Common.CommandProcessor
 		private readonly string _destinationFile;
         private string _srcBackupFile; 
         private string _destBackupFile;
-		#endregion
+	    private readonly bool _failIfExists;
 
-		public RenameFileCommand(string sourceFile, string destinationFile)
+	    #endregion
+
+		public RenameFileCommand(string sourceFile, string destinationFile, bool failIfExists)
 			: base(String.Format("Rename {0} to {1}", sourceFile, destinationFile), true)
 		{
 			Platform.CheckForNullReference(sourceFile, "Source filename");
@@ -58,6 +60,7 @@ namespace ClearCanvas.ImageServer.Common.CommandProcessor
 
 			_sourceFile = sourceFile;
 			_destinationFile = destinationFile;
+		    _failIfExists = failIfExists;
 		}
 
 		protected override void OnExecute()
@@ -65,8 +68,10 @@ namespace ClearCanvas.ImageServer.Common.CommandProcessor
             if (RequiresRollback)
                 Backup();
 
-			if (File.Exists(_destinationFile))
+            if (File.Exists(_destinationFile))
 			{
+                if (_failIfExists)
+                    throw new ApplicationException(String.Format("Destination file already exists: {0}", _destinationFile));
 				File.Delete(_destinationFile);
 			}
